improve Magnet UI architecture
